1.部分button变窄问题； 2.profile界面别人共享kids不显示问题 3.退出登陆处理流程完善 4.event显示新增当前选择kids过滤条件

+++ Swing/app/src/main/java/com/kidsdynamic/swing/domain/DeviceManager.java
@@ -219,7 +219,7 @@ public class DeviceManager {
         kidsDataStore.delKidsInfo(kidsId);
     }
 
-    private static List<DB_Kids> getDBKidsForShared(int suhHostId, long parentId, List<KidInfo> kidInfos){
+    private static List<DB_Kids> getDBKidsForShared(long suhHostId, long parentId, List<KidInfo> kidInfos){
         List<DB_Kids> db_kidsList = new ArrayList<>(kidInfos.size());
 
         for (KidInfo kidInfo : kidInfos) {
